Issue #SB-1905 fix: Review changes
